Begin refactor into class + modules
